New entropy based attribute meta-features
